requirements.txt
pandas
psycopg2-binary
sqlalchemy
psutil
